requirements.txt
numpy 
pandas
pymongo
python-dotenv
scikit-learn
seaborn
catboost
xgboost
# -e .